chore(eggDrop/consolidate): Tests and Problems in one project
Egg Drop taken care of.
