+++ EggDropTrialQuestion/EggDropTrialQuestion.csproj
@@ -30,6 +30,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="nunit.framework, Version=3.9.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.9.0\lib\net45\nunit.framework.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -40,13 +43,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EggDropSolverDynamicProgramming.cs" />
-    <Compile Include="EggDropSolverRecursive.cs" />
-    <Compile Include="IOptimalEggDropTrials.cs" />
-    <Compile Include="MinEggDropTrials.cs" />
+    <Compile Include="Problem\EggDropSolverDynamicProgramming.cs" />
+    <Compile Include="Problem\EggDropSolverRecursive.cs" />
+    <Compile Include="Problem\IOptimalEggDropTrials.cs" />
+    <Compile Include="Problem\MinEggDropTrials.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tests\EggDropTrialTest.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="packages.config" />
     <None Include="README.md" />
   </ItemGroup>
   <ItemGroup>
