--- EggDropTrialQuestion/MinEggDropTrials.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace EggDropTrialQuestion
-{
-    public class MinEggDropTrials
-    {
-        public static int GetMinimumDropTrials()
-        {
-            return 1;
-        }
-    }
-}
