+++ EggDropTrialQuestion/Problem/MinEggDropTrials.cs
@@ -1,0 +1,10 @@
+﻿namespace EggDropTrialQuestion.Problem
+{
+    public class MinEggDropTrials
+    {
+        public static int GetMinimumDropTrials()
+        {
+            return 1;
+        }
+    }
+}
