+++ InterviewQuestions.sln
@@ -61,8 +61,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FirstNonRepeatingCharacters
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EggDropTrialQuestion", "EggDropTrialQuestion\EggDropTrialQuestion.csproj", "{C788EAE6-D9DD-473F-A551-62DFD2EE2689}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EggDropTrialTester", "EggDropTrialTester\EggDropTrialTester.csproj", "{1E686BC1-C10E-4D60-805E-C6F81DED2212}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListMergeSort", "LinkedListMergeSort\LinkedListMergeSort.csproj", "{C6226CB8-76F1-476F-A13E-C0F4F44490FD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListMergeSortTester", "LinkedListMergeSortTester\LinkedListMergeSortTester.csproj", "{3B2C42BA-2278-4709-B32D-8832B85F31DD}"
@@ -195,10 +193,6 @@ Global
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1E686BC1-C10E-4D60-805E-C6F81DED2212}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1E686BC1-C10E-4D60-805E-C6F81DED2212}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1E686BC1-C10E-4D60-805E-C6F81DED2212}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1E686BC1-C10E-4D60-805E-C6F81DED2212}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
